requirements.txt
opencv-python
Pillow
numpy
torch
transformers
paddleocr
paddlepaddle
sentencepiece 
easyocr